refactor(#188): move code to class
