+++ src/lib/effector/filters.ts
@@ -1,0 +1,4 @@
+import { isEqual } from 'lodash';
+
+export const updateNotEqualFilter = <X, Y>(update: X, state: Y) =>
+	!isEqual(update, state);
